Reorganize the ballot_generator and nuke AC model
