+++ tests/data/csv/test_csv_pp_rankings.csv
@@ -7,28 +7,28 @@ Includes Voter Set
 True
 =,=,=,=,=,=,=,=,=,=
 &,Ranking_1,Ranking_2,Ranking_3,&,Weight,&,Voter Set
-&,"{'Alex', 'Alei'}",{},{'C'},&,3/2,&,Chris,Peter
-&,"{'Alex', 'Alei'}",{},{'C'},&,1/2,&,Moon
+&,"{'Alei', 'Alex'}",{},{'C'},&,3/2,&,Chris,Peter
+&,"{'Alei', 'Alex'}",{},{'C'},&,1/2,&,Moon
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
-&,"{'Alex', 'Alei'}",{},{'C'},&,3/2,&,Chris,Peter
-&,"{'Alex', 'Alei'}",{},{'C'},&,1/2,&,Moon
+&,"{'Alei', 'Alex'}",{},{'C'},&,3/2,&,Chris,Peter
+&,"{'Alei', 'Alex'}",{},{'C'},&,1/2,&,Moon
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
-&,"{'Alex', 'Alei'}",{},{'C'},&,3/2,&,Chris,Peter
-&,"{'Alex', 'Alei'}",{},{'C'},&,1/2,&,Moon
+&,"{'Alei', 'Alex'}",{},{'C'},&,3/2,&,Chris,Peter
+&,"{'Alei', 'Alex'}",{},{'C'},&,1/2,&,Moon
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
-&,"{'Alex', 'Alei'}",{},{'C'},&,3/2,&,Chris,Peter
-&,"{'Alex', 'Alei'}",{},{'C'},&,1/2,&,Moon
+&,"{'Alei', 'Alex'}",{},{'C'},&,3/2,&,Chris,Peter
+&,"{'Alei', 'Alex'}",{},{'C'},&,1/2,&,Moon
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
-&,"{'Alex', 'Alei'}",{},{'C'},&,3/2,&,Chris,Peter
-&,"{'Alex', 'Alei'}",{},{'C'},&,1/2,&,Moon
+&,"{'Alei', 'Alex'}",{},{'C'},&,3/2,&,Chris,Peter
+&,"{'Alei', 'Alex'}",{},{'C'},&,1/2,&,Moon
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
 &,{'Alei'},{'Alex'},,&,1/1,&
